chore: 역할에 따른 패키지 분리
