+++ app/src/main/java/com/egobook/app/di/RepositoryModule.kt
@@ -15,8 +15,8 @@ import com.egobook.app.ui.shop.NetworkStoreRepository
 import com.egobook.app.ui.shop.StoreRepository
 import dagger.Binds
 import com.egobook.app.domain.repository.QuestionRepository
-import com.egobook.app.ui.home.NetworkUserRepository
-import com.egobook.app.ui.home.UserRepository
+import com.egobook.app.ui.home.repository.NetworkUserRepository
+import com.egobook.app.ui.home.repository.UserRepository
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
