+++ app/src/main/java/com/egobook/app/ui/home/HomeViewModel.kt
@@ -2,10 +2,11 @@ package com.egobook.app.ui.home
 
 import android.util.Log
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
-import androidx.lifecycle.viewmodel.CreationExtras
-import com.egobook.app.ui.shop.StoreRepository
+import com.egobook.app.ui.home.repository.UserRepository
+import com.egobook.app.ui.home.user.Ink
+import com.egobook.app.ui.home.user.Level
+import com.egobook.app.ui.home.user.User
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
