+++ app/src/main/java/com/egobook/app/ui/home/notification/NotificationStatus.kt
@@ -1,4 +1,4 @@
-package com.egobook.app.ui.home
+package com.egobook.app.ui.home.notification
 
 enum class NotificationStatus {
     READ, UNREAD
