+++ app/src/main/java/com/egobook/app/ui/home/repository/BaseResponse.kt
@@ -1,0 +1,8 @@
+package com.egobook.app.ui.home.repository
+
+data class BaseResponse<T>(
+    val code: String,
+    val message: String,
+    val status: Int,
+    val data: T
+)
