+++ app/src/main/java/com/egobook/app/ui/home/ui/AdDialog.kt
@@ -1,4 +1,4 @@
-package com.egobook.app.ui.home
+package com.egobook.app.ui.home.ui
 
 import android.graphics.Color
 import android.graphics.drawable.ColorDrawable
@@ -7,13 +7,12 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.DialogFragment
-import com.egobook.app.R
-import com.egobook.app.databinding.DialogRadarBinding
+import com.egobook.app.databinding.DialogAdBinding
 import com.egobook.app.removeScreenBlur
 
-class RadarDialog() : DialogFragment() {
+class AdDialog() : DialogFragment() {
 
-    private var _binding: DialogRadarBinding? = null
+    private var _binding: DialogAdBinding? = null
     private val binding get() = checkNotNull(_binding) { "Fragment가 제거되었습니다." }
 
     override fun onCreateView(
@@ -22,15 +21,13 @@ class RadarDialog() : DialogFragment() {
         savedInstanceState: Bundle?
     ): View {
         dialog?.window?.setBackgroundDrawable(ColorDrawable(Color.TRANSPARENT))
-        _binding = DialogRadarBinding.inflate(inflater, container, false)
+        _binding = DialogAdBinding.inflate(inflater, container, false)
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        val radarView: RadarView = view.findViewById(R.id.custom_radar_view)
-        radarView.setRadarData(listOf(4, 5, 2, 1, 3))
-        binding.ivRadarClose.setOnClickListener {
+        binding.btnBack.setOnClickListener {
             removeScreenBlur()
             dismiss()
         }
