+++ app/src/main/java/com/egobook/app/ui/home/user/Level.kt
@@ -1,7 +1,9 @@
-package com.egobook.app.ui.home
+package com.egobook.app.ui.home.user
+
+import com.egobook.app.ui.home.user.LevelType
 
 data class Level(val number: Int) {
-    val type: LevelType = LevelType.of(number)
+    val type: LevelType = LevelType.Companion.of(number)
     init {
         check(number in MINIMUM_LEVEL..MAXIMUM_LEVEL) { "올바른 레벨 범위가 아닙니다." }
     }
