+++ app/src/main/java/com/egobook/app/ui/home/user/User.kt
@@ -1,3 +1,3 @@
-package com.egobook.app.ui.home
+package com.egobook.app.ui.home.user
 
 data class User(val level: Level, val ink: Ink)
