+++ app/src/main/res/layout/activity_main.xml
@@ -78,6 +78,6 @@
         android:layout_width="259dp"
         android:layout_height="match_parent"
         android:layout_gravity="end"
-        android:name="com.egobook.app.ui.home.NotificationFragment"/>
+        android:name="com.egobook.app.ui.home.ui.NotificationFragment"/>
 
 </androidx.drawerlayout.widget.DrawerLayout>
