+++ app/src/main/res/navigation/bottom_navigation.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/menu_home"
-        android:name="com.egobook.app.ui.home.HomeFragment"
+        android:name="com.egobook.app.ui.home.ui.HomeFragment"
         android:label="홈">
 
         <action
